Move policy heads of DDPG and SAC to new module
