+++ rl_blox/algorithm/sac.py
@@ -1,6 +1,5 @@
 from collections import namedtuple
 
-import distrax
 import gymnasium as gym
 import jax
 import jax.numpy as jnp
@@ -9,91 +8,14 @@ import optax
 import tqdm
 from flax import nnx
 
+from ..blox.function_approximator.gaussian_mlp import GaussianMLP
+from ..blox.function_approximator.mlp import MLP
+from ..blox.function_approximator.policy_head import (
+    GaussianTanhPolicy,
+    StochasticPolicyBase,
+)
 from ..logging.logger import LoggerBase
 from .ddpg import ReplayBuffer, mse_action_value_loss, update_target
-from ..blox.function_approximator.mlp import MLP
-from ..blox.function_approximator.gaussian_mlp import GaussianMLP
-
-
-# TODO merge with Gaussian policy from REINFORCE branch
-class StochasticPolicyBase(nnx.Module):
-    """Base class for probabilistic policies."""
-
-    def __call__(self, observation: jnp.ndarray) -> jnp.ndarray:
-        """Compute action probabilities for given observation."""
-        raise NotImplementedError("Subclasses must implement __call__ method.")
-
-    def sample(self, observation: jnp.ndarray, key: jnp.ndarray) -> jnp.ndarray:
-        """Sample action from policy given observation."""
-        raise NotImplementedError("Subclasses must implement sample method.")
-
-    def log_probability(
-        self,
-        observation: jnp.ndarray,
-        action: jnp.ndarray,
-    ) -> jnp.ndarray:
-        """Compute log probability of action given observation."""
-        raise NotImplementedError(
-            "Subclasses must implement log_probability method."
-        )
-
-
-# TODO merge with Gaussian policy from REINFORCE branch
-class GaussianPolicy(StochasticPolicyBase):
-    r"""Gaussian policy represented with a function approximator.
-
-    The gaussian policy maps observations to mean and log variance of an
-    action, hence, represents the distribution :math:`\pi(a|o)`.
-    """
-
-    net: nnx.Module
-    """Underlying function approximator."""
-
-    action_scale: nnx.Variable[jnp.ndarray]
-    """Scales for each component of the action."""
-
-    action_bias: nnx.Variable[jnp.ndarray]
-    """Offset for each component of the action."""
-
-    def __init__(self, policy_net: nnx.Module, action_space: gym.spaces.Box):
-        self.net = policy_net
-        self.action_scale = nnx.Variable(
-            jnp.array((action_space.high - action_space.low) / 2.0)
-        )
-        self.action_bias = nnx.Variable(
-            jnp.array((action_space.high + action_space.low) / 2.0)
-        )
-
-    def __call__(
-        self, observation: jnp.ndarray
-    ) -> tuple[jnp.ndarray, jnp.ndarray]:
-        y, log_var = self.net(observation)
-        mean = nnx.tanh(y) * jnp.broadcast_to(
-            self.action_scale.value, y.shape
-        ) + jnp.broadcast_to(self.action_bias.value, y.shape)
-        log_std = jnp.clip(0.5 * log_var, -20.0, 2.0)
-        std = jnp.exp(log_std)
-        return mean, std
-
-    def sample(self, observation: jnp.ndarray, key: jnp.ndarray) -> jnp.ndarray:
-        """Sample action from Gaussian distribution."""
-        mean, std = self(observation)
-        return jax.random.normal(key, mean.shape) * std + mean
-
-    def log_probability(
-        self,
-        observation: jnp.ndarray,
-        action: jnp.ndarray,
-    ) -> jnp.ndarray:
-        """Compute log probability of action given observation."""
-        mean, std = self(observation)
-        # same as
-        # -jnp.log(std)
-        # - 0.5 * jnp.log(2.0 * jnp.pi)
-        # - 0.5 * ((action - mean) / std) ** 2
-        return distrax.MultivariateNormalDiag(
-            loc=mean, scale_diag=std
-        ).log_prob(action)
 
 
 def sac_actor_loss(
@@ -285,7 +207,7 @@ def create_sac_state(
         policy_activation,
         nnx.Rngs(seed),
     )
-    policy = GaussianPolicy(policy_net, env.action_space)
+    policy = GaussianTanhPolicy(policy_net, env.action_space)
     policy_optimizer = nnx.Optimizer(
         policy, optax.adam(learning_rate=policy_learning_rate)
     )
